Remove usize truncations

+++ src/sstable/poswriter.rs
@@ -3,20 +3,20 @@ use std::io::{Result, Write};
 #[derive(Debug)]
 pub struct PosWriter<W> {
     w: W,
-    offset: usize,
+    offset: u64,
 }
 
 impl<W> PosWriter<W> {
-    pub fn new(w: W, offset: usize) -> Self {
+    pub fn new(w: W, offset: u64) -> Self {
         PosWriter {
             w,
             offset,
         }
     }
-    pub fn current_offset(&self) -> usize {
+    pub fn current_offset(&self) -> u64 {
         self.offset
     }
-    pub fn reset_offset(&mut self, offset: usize) {
+    pub fn reset_offset(&mut self, offset: u64) {
         self.offset = offset;
     }
     pub fn get_mut(&mut self) -> &mut W {
@@ -30,7 +30,7 @@ impl<W> PosWriter<W> {
 impl<W: Write> Write for PosWriter<W> {
     fn write(&mut self, buf: &[u8]) -> Result<usize> {
         let l = self.w.write(buf)?;
-        self.offset += l;
+        self.offset += l as u64;
         Ok(l)
     }
